[♻️refactor]: enum 위치 이동
- types 디렉토리 내에 enum 디렉토리 만들어서 넣어둠
